// backend/src/modules/zakat/dto/zakat-result.dto.ts
export class ZakatResultDto {
  id?:             string;
  totalZakatable:  number;
  nisabThreshold:  number;
  isLiable:        boolean;
  zakatAmount:     number;
  breakdown: {
    gold:          number;
    silver:        number;
    cash:          number;
    business:      number;
    debtDeduction: number;
  };
  createdAt?: string;
}